chore: remove paginated tokens flag
