+++ src/authorizations/components/TokenList.tsx
@@ -1,10 +1,10 @@
 // Libraries
-import React, {PureComponent} from 'react'
+import React, {PureComponent, RefObject} from 'react'
 import memoizeOne from 'memoize-one'
 import isEqual from 'lodash/isEqual'
 
 // Components
-import {Overlay, ResourceList} from '@influxdata/clockface'
+import {Overlay, PaginationNav, ResourceList} from '@influxdata/clockface'
 import {TokenRow} from 'src/authorizations/components/TokenRow'
 import EditTokenOverlay from 'src/authorizations/components/EditTokenOverlay'
 
@@ -21,10 +21,13 @@ type SortKey = keyof Authorization
 interface Props {
   auths: Authorization[]
   emptyState: JSX.Element
+  pageHeight: number
+  pageWidth: number
   searchTerm: string
   sortKey: string
   sortDirection: Sort
   sortType: SortTypes
+  tokenCount: number
   onClickColumn: (nextSort: Sort, sortKey: SortKey) => void
 }
 
@@ -33,10 +36,15 @@ interface State {
   authInView: Authorization
 }
 
-export class TokenList extends PureComponent<Props, State> {
+export default class TokenList extends PureComponent<Props, State> {
   private memGetSortedResources = memoizeOne<typeof getSortedResources>(
     getSortedResources
   )
+
+  private paginationRef: RefObject<HTMLDivElement>
+  public currentPage: number = 1
+  public rowsPerPage: number = 10
+  public totalPages: number
 
   constructor(props) {
     super(props)
@@ -46,7 +54,25 @@ export class TokenList extends PureComponent<Props, State> {
     }
   }
 
+  public componentDidMount() {
+    const params = new URLSearchParams(window.location.search)
+    const urlPageNumber = parseInt(params.get('page'), 10)
+
+    const passedInPageIsValid =
+      urlPageNumber && urlPageNumber <= this.totalPages && urlPageNumber > 0
+
+    if (passedInPageIsValid) {
+      this.currentPage = urlPageNumber
+    }
+  }
+
   public componentDidUpdate(prevProps) {
+    // if the user filters the list while on a page that is
+    // outside the new filtered list put them on the last page of the new list
+    if (this.currentPage > this.totalPages) {
+      this.paginate(this.totalPages)
+    }
+
     const {auths: prevAuths} = prevProps
     const {auths: nextAuths} = this.props
 
@@ -60,18 +86,34 @@ export class TokenList extends PureComponent<Props, State> {
 
   public render() {
     const {isTokenOverlayVisible, authInView} = this.state
+    this.totalPages = Math.max(
+      Math.ceil(this.props.auths.length / this.rowsPerPage),
+      1
+    )
 
     return (
       <>
-        <ResourceList>
+        <ResourceList style={{width: this.props.pageWidth}}>
           <ResourceList.Body
             emptyState={this.props.emptyState}
+            style={{
+              maxHeight: this.props.pageHeight,
+              minHeight: this.props.pageHeight,
+              overflow: 'auto',
+            }}
             testID="token-list"
           >
             {this.rows}
           </ResourceList.Body>
         </ResourceList>
-
+        <PaginationNav.PaginationNav
+          ref={this.paginationRef}
+          style={{width: this.props.pageWidth}}
+          totalPages={this.totalPages}
+          currentPage={this.currentPage}
+          pageRangeOffset={1}
+          onChange={this.paginate}
+        />
         <Overlay visible={isTokenOverlayVisible}>
           <EditTokenOverlay
             auth={authInView}
@@ -80,6 +122,14 @@ export class TokenList extends PureComponent<Props, State> {
         </Overlay>
       </>
     )
+  }
+
+  public paginate = page => {
+    this.currentPage = page
+    const url = new URL(location.href)
+    url.searchParams.set('page', page)
+    history.replaceState(null, '', url.toString())
+    this.forceUpdate()
   }
 
   private get rows(): JSX.Element[] {
@@ -91,13 +141,28 @@ export class TokenList extends PureComponent<Props, State> {
       sortType
     )
 
-    return sortedAuths.map(auth => (
-      <TokenRow
-        key={auth.id}
-        auth={auth}
-        onClickDescription={this.handleClickDescription}
-      />
-    ))
+    const startIndex = this.rowsPerPage * Math.max(this.currentPage - 1, 0)
+    const endIndex = Math.min(
+      startIndex + this.rowsPerPage,
+      this.props.tokenCount
+    )
+
+    const paginatedAuths = []
+    for (let i = startIndex; i < endIndex; i++) {
+      const auth = sortedAuths[i]
+
+      if (auth) {
+        paginatedAuths.push(
+          <TokenRow
+            key={auth.id}
+            auth={auth}
+            onClickDescription={this.handleClickDescription}
+          />
+        )
+      }
+    }
+
+    return paginatedAuths
   }
 
   private handleDismissOverlay = () => {
@@ -105,7 +170,7 @@ export class TokenList extends PureComponent<Props, State> {
   }
 
   private handleClickDescription = (authID: string): void => {
-    const authInView = this.props.auths.find(auth => auth.id === authID)
+    const authInView = this.props.auths.find(a => a.id === authID)
     this.setState({isTokenOverlayVisible: true, authInView})
   }
 }
