+++ src/authorizations/components/TokensTab.tsx
@@ -1,36 +1,35 @@
 // Libraries
-import React, {PureComponent} from 'react'
+import React, {createRef, PureComponent, RefObject} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, RouteComponentProps} from 'react-router-dom'
 import {isEmpty} from 'lodash'
+import {AutoSizer} from 'react-virtualized'
 
 // Components
 import {
   Sort,
   ComponentSize,
   EmptyState,
-  FlexBox,
-  FlexDirection,
-  AlignItems,
   BannerPanel,
   Gradients,
   IconFont,
   InfluxColors,
 } from '@influxdata/clockface'
 import SearchWidget from 'src/shared/components/search_widget/SearchWidget'
-import {TokenList} from 'src/authorizations/components/TokenList'
+import TokenList from 'src/authorizations/components/TokenList'
 import FilterList from 'src/shared/components/FilterList'
 import TabbedPageHeader from 'src/shared/components/tabbed_page/TabbedPageHeader'
 import GenerateTokenDropdown from 'src/authorizations/components/GenerateTokenDropdown'
 import ResourceSortDropdown from 'src/shared/components/resource_sort_dropdown/ResourceSortDropdown'
+import PostDeploymentTokensBanner from 'src/authorizations/components/PostDeploymentTokensBanner'
 
 // Types
 import {AppState, Authorization, ResourceType} from 'src/types'
 import {SortTypes} from 'src/shared/utils/sort'
+import {AuthorizationSortKey} from 'src/shared/components/resource_sort_dropdown/generateSortItems'
 
 // Selectors
 import {getAll} from 'src/resources/selectors'
-import PostDeploymentTokensBanner from './PostDeploymentTokensBanner'
 
 enum AuthSearchKeys {
   Description = 'description',
@@ -49,6 +48,10 @@ interface StateProps {
   tokens: Authorization[]
 }
 
+const DEFAULT_PAGINATION_CONTROL_HEIGHT = 62
+const DEFAULT_TAB_NAVIGATION_HEIGHT = 62
+const DEFAULT_ALERT_HEIGHT = 100
+
 type SortKey = keyof Authorization
 
 type Props = StateProps & RouteComponentProps<{orgID: string}>
@@ -56,6 +59,8 @@ type Props = StateProps & RouteComponentProps<{orgID: string}>
 const FilterAuthorizations = FilterList<Authorization>()
 
 class TokensTab extends PureComponent<Props, State> {
+  private paginationRef: RefObject<HTMLDivElement>
+
   constructor(props) {
     super(props)
     this.state = {
@@ -64,6 +69,35 @@ class TokensTab extends PureComponent<Props, State> {
       sortDirection: Sort.Ascending,
       sortType: SortTypes.String,
     }
+    this.paginationRef = createRef<HTMLDivElement>()
+  }
+
+  public componentDidMount() {
+    const params = new URLSearchParams(window.location.search)
+
+    let sortType: SortTypes = this.state.sortType
+    let sortKey: AuthorizationSortKey = 'description'
+    if (params.get('sortKey') === 'status') {
+      sortKey = 'status'
+      sortType = SortTypes.String
+    } else if (params.get('sortKey') === 'createdAt') {
+      sortKey = 'createdAt'
+      sortType = SortTypes.String
+    }
+
+    let sortDirection: Sort = this.state.sortDirection
+    if (params.get('sortDirection') === Sort.Ascending) {
+      sortDirection = Sort.Ascending
+    } else if (params.get('sortDirection') === Sort.Descending) {
+      sortDirection = Sort.Descending
+    }
+
+    let searchTerm: string = ''
+    if (params.get('searchTerm') !== null) {
+      searchTerm = params.get('searchTerm')
+    }
+
+    this.setState({sortKey, sortDirection, sortType, searchTerm})
   }
 
   public render() {
@@ -74,8 +108,8 @@ class TokensTab extends PureComponent<Props, State> {
       <>
         <SearchWidget
           searchTerm={searchTerm}
-          placeholderText="Filter API Tokens..."
-          onSearch={this.handleChangeSearchTerm}
+          placeholderText="Filter Tokens..."
+          onSearch={this.handleFilterUpdate}
           testID="input-field--filter"
         />
         <ResourceSortDropdown
@@ -91,43 +125,69 @@ class TokensTab extends PureComponent<Props, State> {
 
     const rightHeaderItems = <GenerateTokenDropdown />
 
+    const tokensBanner = () => {
+      return (
+        <>
+          <BannerPanel
+            size={ComponentSize.ExtraSmall}
+            gradient={Gradients.PolarExpress}
+            icon={IconFont.Bell}
+            hideMobileIcon={true}
+            textColor={InfluxColors.Yeti}
+          >
+            <PostDeploymentTokensBanner />
+          </BannerPanel>
+        </>
+      )
+    }
+
     return (
-      <FlexBox
-        direction={FlexDirection.Column}
-        alignItems={AlignItems.Center}
-        margin={ComponentSize.Large}
-      >
-        <BannerPanel
-          size={ComponentSize.ExtraSmall}
-          gradient={Gradients.PolarExpress}
-          icon={IconFont.Bell}
-          hideMobileIcon={true}
-          textColor={InfluxColors.Yeti}
-        >
-          <PostDeploymentTokensBanner />
-        </BannerPanel>
-        <TabbedPageHeader
-          childrenLeft={leftHeaderItems}
-          childrenRight={rightHeaderItems}
-        />
-        <FilterAuthorizations
-          list={tokens}
-          searchTerm={searchTerm}
-          searchKeys={this.searchKeys}
-        >
-          {filteredAuths => (
-            <TokenList
-              auths={filteredAuths}
-              emptyState={this.emptyState}
-              searchTerm={searchTerm}
-              sortKey={sortKey}
-              sortDirection={sortDirection}
-              sortType={sortType}
-              onClickColumn={this.handleClickColumn}
-            />
-          )}
-        </FilterAuthorizations>
-      </FlexBox>
+      <>
+        {tokensBanner()}
+        <AutoSizer>
+          {({width, height}) => {
+            const heightWithPagination =
+              this.paginationRef?.current?.clientHeight +
+                DEFAULT_TAB_NAVIGATION_HEIGHT ||
+              DEFAULT_PAGINATION_CONTROL_HEIGHT +
+                DEFAULT_TAB_NAVIGATION_HEIGHT +
+                DEFAULT_ALERT_HEIGHT
+
+            const adjustedHeight = height - heightWithPagination
+            return (
+              <>
+                <div style={{margin: '10px 0px'}}>
+                  <TabbedPageHeader
+                    childrenLeft={leftHeaderItems}
+                    childrenRight={rightHeaderItems}
+                    width={width}
+                  />
+                  <FilterAuthorizations
+                    list={tokens}
+                    searchTerm={searchTerm}
+                    searchKeys={this.searchKeys}
+                  >
+                    {filteredAuths => (
+                      <TokenList
+                        tokenCount={tokens.length}
+                        auths={filteredAuths}
+                        emptyState={this.emptyState}
+                        pageWidth={width}
+                        pageHeight={adjustedHeight}
+                        searchTerm={searchTerm}
+                        sortKey={sortKey}
+                        sortDirection={sortDirection}
+                        sortType={sortType}
+                        onClickColumn={this.handleClickColumn}
+                      />
+                    )}
+                  </FilterAuthorizations>
+                </div>
+              </>
+            )
+          }}
+        </AutoSizer>
+      </>
     )
   }
 
@@ -136,6 +196,11 @@ class TokensTab extends PureComponent<Props, State> {
     sortDirection: Sort,
     sortType: SortTypes
   ): void => {
+    const url = new URL(location.href)
+    url.searchParams.set('sortKey', sortKey)
+    url.searchParams.set('sortDirection', sortDirection)
+    history.replaceState(null, '', url.toString())
+
     this.setState({sortKey, sortDirection, sortType})
   }
 
@@ -144,7 +209,10 @@ class TokensTab extends PureComponent<Props, State> {
     this.setState({sortKey, sortDirection: nextSort, sortType})
   }
 
-  private handleChangeSearchTerm = (searchTerm: string): void => {
+  private handleFilterUpdate = (searchTerm: string): void => {
+    const url = new URL(location.href)
+    url.searchParams.set('searchTerm', searchTerm)
+    history.replaceState(null, '', url.toString())
     this.setState({searchTerm})
   }
 
@@ -163,7 +231,7 @@ class TokensTab extends PureComponent<Props, State> {
       return (
         <EmptyState size={ComponentSize.Large}>
           <EmptyState.Text>
-            Looks like there aren't any <b>API Tokens</b>, why not generate one?
+            Looks like there aren't any <b>Tokens</b>, why not generate one?
           </EmptyState.Text>
           <GenerateTokenDropdown />
         </EmptyState>
@@ -172,7 +240,7 @@ class TokensTab extends PureComponent<Props, State> {
 
     return (
       <EmptyState size={ComponentSize.Large}>
-        <EmptyState.Text>No API Tokens match your query</EmptyState.Text>
+        <EmptyState.Text>No Tokens match your query</EmptyState.Text>
       </EmptyState>
     )
   }
