+++ src/shared/containers/SetOrg.tsx
@@ -20,7 +20,6 @@ import {
   FlowPage,
   BucketsIndex,
   BucketsIndexPaginated,
-  PaginatedTokensIndex,
   TokensIndex,
   TelegrafsPage,
   ScrapersIndex,
@@ -112,20 +111,6 @@ const SetOrg: FC = () => {
 
   const orgPath = '/orgs/:orgID'
 
-  let CorrectedTokensIndex = (
-    <Route path={`${orgPath}/${LOAD_DATA}/${TOKENS}`} component={TokensIndex} />
-  )
-
-  // paginatedTokens already handles tokensUIRedesign being flagged on
-  if (isFlagEnabled('paginatedTokens')) {
-    CorrectedTokensIndex = (
-      <Route
-        path={`${orgPath}/${LOAD_DATA}/${TOKENS}`}
-        component={PaginatedTokensIndex}
-      />
-    )
-  }
-
   return (
     <PageSpinner loading={loading}>
       <Suspense fallback={<PageSpinner />}>
@@ -213,7 +198,10 @@ const SetOrg: FC = () => {
             path={`${orgPath}/${LOAD_DATA}/${TELEGRAFS}`}
             component={TelegrafsPage}
           />
-          {CorrectedTokensIndex}
+          <Route
+            path={`${orgPath}/${LOAD_DATA}/${TOKENS}`}
+            component={TokensIndex}
+          />
           {isFlagEnabled('fetchAllBuckets') ? (
             <Route
               path={`${orgPath}/${LOAD_DATA}/${BUCKETS}`}
